Add files via upload
The attached SAS code documents that were created to compare worldwide population growth from 1930 to 1970 and its effect on the rainfall in the Panama Canal. Worldwide population data was gathered from the United States census and rainfall data was gathered from the Smithsonian Institution's weather station on Barro Colorado Island, in the middle of the Panama Canal.